gpu: lift the readback availability barrier into a shared header (#3249)
The render backend's `record_host_read_barrier` / `host_read_barrier_for` /
`backend_host_read_barrier_count` move verbatim from `tests/fixtures/render_runner.h`
into `src/gpu/execute/host_read_barrier.hpp` so the live compute backend can call
the same helper instead of growing a second spelling of the rule. `render_runner.h`
re-exports them into `prosper::test`, so every existing call site and the
`host_read_barrier` ctest case are unchanged.

The invalidate half stays in `render_runner.h`: it belongs with the one allocator
in the tree that can hand out non-coherent HOST_CACHED memory.

Pure move plus the re-export; no behaviour change.
